fixed compilation errors

+++ Final Project/main.cpp
@@ -39,7 +39,7 @@ int main(int argc, char *argv[])
 	bf(0) = 0;
 
 	Eigen::VectorXd weights(num_vertices);
-	Eigen::VectorXd mu(num_vertices);
+	Eigen::VectorXi mu(num_vertices);
 	for (int i = 0; i < num_vertices; i++) {
 		weights(i) = 1;
 		mu(i) = 1;
@@ -57,7 +57,7 @@ int main(int argc, char *argv[])
 	Eigen::Vector3d view;
 	// view = (Eigen::Vector3d) camera_pos;
 	view << (double) camera_pos[0], (double) camera_pos[1], (double) camera_pos[2];
-
+	
 	appearance_mimicking_surfaces(V, F, view, lambda_min, lambda_max, bf, weights, mu, DV);
 
 	std::cout << R"(
